+++ front/src/views/Hasiera/Hasiera.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import ReactMarkdown from 'react-markdown';
 
 import Gezia from '../../assets/gezia.svg';
-import { Gainburua } from '../../components/Gainburua';
 import { Oina } from '../../components/Oina';
 import { HasieraContent } from '../../domain/hasiera/HasieraContent';
 import { Container } from '../components/Container';
+import { Gainburua } from '../components/Gainburua';
 import * as styles from './Hasiera.module.scss';
 
 interface Props {
