+++ front/src/views/components/Gainburua/AtalaName.ts
