+++ front/src/views/components/Gainburua/Gainburua.module.scss
@@ -1,6 +1,6 @@
-@import '../../styles/tools/mediaQueries.scss';
-@import '../../styles/tools/mixins/font.scss';
-@import '../../styles/tools/mixins/container.scss';
+@import '../../../styles/tools/mediaQueries.scss';
+@import '../../../styles/tools/mixins/font.scss';
+@import '../../../styles/tools/mixins/container.scss';
 
 .geziaWrapper {
   display: none;
