+++ front/src/views/components/Gainburua/components/MugikorNabigazioa/MugikorNabigazioa.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/tools/mixins/font.scss';
+@import '../../../../../styles/tools/mixins/font.scss';
 
 .kontaktuaWrapper {
   display: flex;
