+++ front/src/views/components/Gainburua/index.ts
